improve resource table
